# source/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rejintech API",
    "description": "Rejintech 프로젝트 REST API 문서",
    "version": "1.0.0",
    "contact": {
      "name": "Rejintech Support",
      "email": "admin@rejintech.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "개발 서버"
    },
    {
      "url": "https://api.rejintech.com",
      "description": "프로덕션 서버"
    }
  ],
  "paths": {
...
    "/api/procurement/statistics/institutions": {
      "get": {
        "tags": ["조달청"],
        "summary": "수요기관별 통계 조회",
        "description": "수요기관별 납품요구 통계를 조회합니다",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "통계 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/procurement/statistics/companies": {
      "get": {
        "tags": ["조달청"],
        "summary": "업체별 통계 조회",
        "description": "업체별 납품요구 통계를 조회합니다",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "통계 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/procurement/statistics/products": {
      "get": {
        "tags": ["조달청"],
        "summary": "품목별 통계 조회",
        "description": "품목별 납품요구 통계를 조회합니다",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "통계 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/procurement/filter-options": {
      "get": {
        "tags": ["조달청"],
        "summary": "필터 옵션 조회",
        "description": "필터링에 사용할 수 있는 옵션들을 조회합니다",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "필터 옵션 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterOptionsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer 토큰을 사용한 인증"
      }
    },
    "schemas": {
...
      "StatisticsResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": true},
          "message": {"type": "string", "example": "통계 조회 성공"},
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {"type": "string"},
                "delivery_count": {"type": "integer"},
                "total_amount": {"type": "number"},
                "avg_amount": {"type": "number"}
              }
            }
          },
          "timestamp": {"type": "string", "format": "date-time"}
        }
      },
      "FilterOptionsResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": true},
          "message": {"type": "string", "example": "필터 옵션 조회 성공"},
          "data": {
            "type": "object",
            "properties": {
              "filterPrdctClsfcNoNm": {
                "type": "array",
                "items": {"type": "string"}
              },
              "filterDminsttNm": {
                "type": "array",
                "items": {"type": "string"}
              },
              "filterCorpNm": {
                "type": "array",
                "items": {"type": "string"}
              }
            }
          },
          "timestamp": {"type": "string", "format": "date-time"}
        }
      }
    }
  },
...
} 